fix: move permission service to one package * remove unused basePermissionService methods * add todos for permission checks * reorganize imports
