+++ shogun-boot/src/test/java/de/terrestris/shogun/boot/controller/BaseControllerTest.java
@@ -29,8 +29,8 @@ import de.terrestris.shogun.lib.repository.BaseCrudRepository;
 import de.terrestris.shogun.lib.repository.UserRepository;
 import de.terrestris.shogun.lib.repository.security.permission.UserClassPermissionRepository;
 import de.terrestris.shogun.lib.repository.security.permission.UserInstancePermissionRepository;
-import de.terrestris.shogun.lib.service.security.permission.internal.UserClassPermissionService;
-import de.terrestris.shogun.lib.service.security.permission.internal.UserInstancePermissionService;
+import de.terrestris.shogun.lib.service.security.permission.UserClassPermissionService;
+import de.terrestris.shogun.lib.service.security.permission.UserInstancePermissionService;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
