+++ shogun-lib/src/main/java/de/terrestris/shogun/lib/security/access/entity/BaseEntityPermissionEvaluator.java
@@ -21,10 +21,10 @@ import de.terrestris.shogun.lib.model.BaseEntity;
 import de.terrestris.shogun.lib.model.User;
 import de.terrestris.shogun.lib.model.security.permission.PermissionCollection;
 import de.terrestris.shogun.lib.repository.BaseCrudRepository;
-import de.terrestris.shogun.lib.service.security.permission.internal.GroupClassPermissionService;
-import de.terrestris.shogun.lib.service.security.permission.internal.GroupInstancePermissionService;
-import de.terrestris.shogun.lib.service.security.permission.internal.UserClassPermissionService;
-import de.terrestris.shogun.lib.service.security.permission.internal.UserInstancePermissionService;
+import de.terrestris.shogun.lib.service.security.permission.GroupClassPermissionService;
+import de.terrestris.shogun.lib.service.security.permission.GroupInstancePermissionService;
+import de.terrestris.shogun.lib.service.security.permission.UserClassPermissionService;
+import de.terrestris.shogun.lib.service.security.permission.UserInstancePermissionService;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.aop.framework.AopProxyUtils;
 import org.springframework.beans.factory.annotation.Autowired;
