+++ shogun-lib/src/main/java/de/terrestris/shogun/lib/service/security/permission/GroupClassPermissionServiceSecured.java
@@ -22,15 +22,9 @@ import de.terrestris.shogun.lib.model.Group;
 import de.terrestris.shogun.lib.model.User;
 import de.terrestris.shogun.lib.model.security.permission.GroupClassPermission;
 import de.terrestris.shogun.lib.model.security.permission.PermissionCollection;
-import de.terrestris.shogun.lib.service.security.permission.internal.GroupClassPermissionService;
-import lombok.extern.log4j.Log4j2;
-import org.springframework.data.jpa.domain.Specification;
-import org.springframework.security.access.prepost.PostAuthorize;
-import org.springframework.security.access.prepost.PostFilter;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
@@ -93,7 +87,7 @@ public class GroupClassPermissionServiceSecured extends GroupClassPermissionServ
     }
 
     @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PreAuthorize("hasRole('ROLE_ADMIN') or hasPermission(#clazz, 'READ')")
     // todo: how to check permission for this?
     public void setPermission(Class<? extends BaseEntity> clazz, Group group, PermissionCollectionType permissionCollectionType) {
         super.setPermission(clazz, group, permissionCollectionType);
@@ -111,42 +105,4 @@ public class GroupClassPermissionServiceSecured extends GroupClassPermissionServ
         super.deleteFor(persistedEntity, group);
     }
 
-    // basePermissionService methods
-    // todo: add permissions for non-admins
-
-    @Override
-    @PostFilter("hasRole('ROLE_ADMIN')")
-    public List<GroupClassPermission> findAll() {
-        return super.findAll();
-    }
-
-    @Override
-    @PostFilter("hasRole('ROLE_ADMIN')")
-    public List<GroupClassPermission> findAllBy(Specification specification) {
-        return super.findAllBy(specification);
-    }
-
-    @Override
-    @PostAuthorize("hasRole('ROLE_ADMIN')")
-    public Optional<GroupClassPermission> findOne(Long id) {
-        return super.findOne(id);
-    }
-
-    @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public GroupClassPermission create(GroupClassPermission entity) {
-        return super.create(entity);
-    }
-
-    @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public GroupClassPermission update(Long id, GroupClassPermission entity) throws IOException {
-        return super.update(id, entity);
-    }
-
-    @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public void delete(GroupClassPermission entity) {
-        super.delete(entity);
-    }
 }
