+++ shogun-lib/src/main/java/de/terrestris/shogun/lib/service/security/permission/UserClassPermissionServiceSecured.java
@@ -19,18 +19,11 @@ package de.terrestris.shogun.lib.service.security.permission;
 import de.terrestris.shogun.lib.enumeration.PermissionCollectionType;
 import de.terrestris.shogun.lib.model.BaseEntity;
 import de.terrestris.shogun.lib.model.User;
-import de.terrestris.shogun.lib.model.security.permission.GroupInstancePermission;
 import de.terrestris.shogun.lib.model.security.permission.PermissionCollection;
 import de.terrestris.shogun.lib.model.security.permission.UserClassPermission;
-import de.terrestris.shogun.lib.service.security.permission.internal.UserClassPermissionService;
-import lombok.extern.log4j.Log4j2;
-import org.springframework.data.jpa.domain.Specification;
-import org.springframework.security.access.prepost.PostAuthorize;
-import org.springframework.security.access.prepost.PostFilter;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
@@ -97,39 +90,4 @@ public class UserClassPermissionServiceSecured extends UserClassPermissionServic
     // basePermissionService methods
     // todo: add permissions for non-admins
 
-    @Override
-    @PostFilter("hasRole('ROLE_ADMIN')")
-    public List<UserClassPermission> findAll() {
-        return super.findAll();
-    }
-
-    @Override
-    @PostFilter("hasRole('ROLE_ADMIN')")
-    public List<UserClassPermission> findAllBy(Specification specification) {
-        return super.findAllBy(specification);
-    }
-
-    @Override
-    @PostAuthorize("hasRole('ROLE_ADMIN')")
-    public Optional<UserClassPermission> findOne(Long id) {
-        return super.findOne(id);
-    }
-
-    @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public UserClassPermission create(UserClassPermission entity) {
-        return super.create(entity);
-    }
-
-    @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public UserClassPermission update(Long id, UserClassPermission entity) throws IOException {
-        return super.update(id, entity);
-    }
-
-    @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public void delete(UserClassPermission entity) {
-        super.delete(entity);
-    }
 }
