+++ shogun-lib/src/main/java/de/terrestris/shogun/lib/service/security/permission/UserInstancePermissionServiceSecured.java
@@ -21,14 +21,9 @@ import de.terrestris.shogun.lib.model.BaseEntity;
 import de.terrestris.shogun.lib.model.User;
 import de.terrestris.shogun.lib.model.security.permission.PermissionCollection;
 import de.terrestris.shogun.lib.model.security.permission.UserInstancePermission;
-import de.terrestris.shogun.lib.service.security.permission.internal.UserInstancePermissionService;
-import org.springframework.data.jpa.domain.Specification;
-import org.springframework.security.access.prepost.PostAuthorize;
-import org.springframework.security.access.prepost.PostFilter;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
@@ -38,6 +33,7 @@ public class UserInstancePermissionServiceSecured extends UserInstancePermission
 
     @Override
     @PreAuthorize("hasRole('ROLE_ADMIN') or hasPermission(#user, 'READ')")
+    // todo: postfilter permission check
     public List<UserInstancePermission> findFor(User user) {
         return super.findFor(user);
     }
@@ -98,42 +94,4 @@ public class UserInstancePermissionServiceSecured extends UserInstancePermission
         super.deleteFor(persistedEntity, user);
     }
 
-    // basePermissionService methods
-    // todo: add permissions for non-admins
-
-    @Override
-    @PostFilter("hasRole('ROLE_ADMIN')")
-    public List<UserInstancePermission> findAll() {
-        return super.findAll();
-    }
-
-    @Override
-    @PostFilter("hasRole('ROLE_ADMIN')")
-    public List<UserInstancePermission> findAllBy(Specification specification) {
-        return super.findAllBy(specification);
-    }
-
-    @Override
-    @PostAuthorize("hasRole('ROLE_ADMIN')")
-    public Optional<UserInstancePermission> findOne(Long id) {
-        return super.findOne(id);
-    }
-
-    @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public UserInstancePermission create(UserInstancePermission entity) {
-        return super.create(entity);
-    }
-
-    @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public UserInstancePermission update(Long id, UserInstancePermission entity) throws IOException {
-        return super.update(id, entity);
-    }
-
-    @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public void delete(UserInstancePermission entity) {
-        super.delete(entity);
-    }
 }
