+++ shogun-lib/src/test/java/de/terrestris/shogun/lib/security/access/entity/DefaultPermissionEvaluatorTest.java
@@ -21,10 +21,10 @@ import de.terrestris.shogun.lib.enumeration.PermissionType;
 import de.terrestris.shogun.lib.model.BaseEntity;
 import de.terrestris.shogun.lib.model.User;
 import de.terrestris.shogun.lib.model.security.permission.PermissionCollection;
-import de.terrestris.shogun.lib.service.security.permission.internal.GroupClassPermissionService;
-import de.terrestris.shogun.lib.service.security.permission.internal.GroupInstancePermissionService;
-import de.terrestris.shogun.lib.service.security.permission.internal.UserClassPermissionService;
-import de.terrestris.shogun.lib.service.security.permission.internal.UserInstancePermissionService;
+import de.terrestris.shogun.lib.service.security.permission.GroupClassPermissionService;
+import de.terrestris.shogun.lib.service.security.permission.GroupInstancePermissionService;
+import de.terrestris.shogun.lib.service.security.permission.UserClassPermissionService;
+import de.terrestris.shogun.lib.service.security.permission.UserInstancePermissionService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
